RASS-cli enhancements
Move deployment into its own module.

Add command for managing robot fire wall.

Move ssh into its own module.

Add auto host selection for single host robots.
