+++ packages/rass-cli/shell.nix
@@ -1,6 +1,9 @@
-{ pkgs ? import <nixpkgs> {}, ... }:
+{ pkgs ? import <nixpkgs> { }, ... }:
+let
+  rust = import ../../nix/rust.nix { pkgs = pkgs; };
+in
 pkgs.mkShell {
   buildInputs = [
-    (import ./. { pkgs = pkgs; })
+    rust
   ];
 }
